<feat> : <article summary api 구현 및 directory 구조 변경> [지라 이슈 링크]
