+++ src/main/java/droni/backend/api/droniuser/exception/DroniUserException.java
@@ -1,0 +1,10 @@
+package droni.backend.api.droniuser.exception;
+
+import droni.backend.global.exception.BaseException;
+import org.springframework.http.HttpStatus;
+
+public class DroniUserException extends BaseException {
+    public DroniUserException(HttpStatus httpStatus, String msg) {
+        super(httpStatus, msg);
+    }
+}
