+++ src/main/java/droni/backend/api/expert/entity/ExpertReview.java
@@ -1,0 +1,27 @@
+package droni.backend.api.expert.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Builder
+@Table(name = "expert_review")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class ExpertReview {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long reviewId;
+    @ManyToOne
+    @JoinColumn(name = "expert_id", nullable = false)
+    private DroniExpert expert;
+    private Integer score;
+    private String content;
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+    private LocalDateTime deletedAt;
+}
