+++ src/main/java/droni/backend/api/insectcontrol/dto/InsectBidStatus.java
@@ -1,0 +1,10 @@
+package droni.backend.api.insectcontrol.dto;
+
+import lombok.Getter;
+
+@Getter
+public enum InsectBidStatus {
+    REQUESTING,
+    RESERVED,
+    COMPLETED
+}
