+++ src/main/java/droni/backend/api/message/dto/DroniServiceKind.java
@@ -1,0 +1,8 @@
+package droni.backend.api.message.dto;
+
+import lombok.Getter;
+
+@Getter
+public enum DroniServiceKind {
+    INSECTCONTROL
+}
