+++ src/main/java/droni/backend/config/security/SecurityConfig.java
@@ -35,6 +35,7 @@ public class SecurityConfig {
             antMatcher("/reissue"),
             antMatcher("/favicon.ico"),
             antMatcher("/swagger-ui/**"),
+            antMatcher("/swagger-ui.html"),
             antMatcher("/v3/api-docs/**")
     };
