--- src/main/java/droni/backend/expert/entity/DroniExpert.java
@@ -1,37 +1,0 @@
-package droni.backend.expert.entity;
-
-import droni.backend.droniuser.entity.DroniUser;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Entity
-@Builder
-@Getter
-@Table(name = "expert")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class DroniExpert {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long expertId;
-    @OneToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private DroniUser user;
-    @Builder.Default
-    private Integer completeRequest = 0;
-    private String introduction;
-    @OneToOne
-    @JoinColumn(name = "license_id")
-    private BusinessLicense license;
-    @OneToMany(mappedBy = "expert")
-    private List<ExpertCareer> careers;
-    @OneToMany(mappedBy = "expert")
-    private List<ExpertReview> reviews;
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-    private LocalDateTime deletedAt;
-
-}
