--- src/main/java/droni/backend/insectcontrol/dto/InsectBidStatus.java
@@ -1,10 +1,0 @@
-package droni.backend.insectcontrol.dto;
-
-import lombok.Getter;
-
-@Getter
-public enum InsectBidStatus {
-    REQUESTING,
-    RESERVED,
-    COMPLETED
-}
