--- src/main/java/droni/backend/message/dto/DroniServiceKind.java
@@ -1,8 +1,0 @@
-package droni.backend.message.dto;
-
-import lombok.Getter;
-
-@Getter
-public enum DroniServiceKind {
-    INSECTCONTROL
-}
