--- src/main/java/droni/backend/test/TestController.java
@@ -1,23 +1,0 @@
-package droni.backend.test;
-
-import backend.generated_api.TestApi;
-import droni.backend.oauth2.execption.JwtExpiredException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class TestController implements TestApi {
-
-
-    @Override
-    public ResponseEntity<String> expireTestGet() {
-        throw new JwtExpiredException("JWT is expired");
-    }
-
-    @Override
-    public ResponseEntity<String> sucessTestGet() {
-        return ResponseEntity.ok("Successfully returned");
-    }
-}
