+++ src/main/resources/openapi.yaml
@@ -2,6 +2,8 @@ openapi: 3.1.0
 servers:
   - description: localhost
     url: http://localhost:8080
+  - description: dev(ncp)
+    url: http://223.130.147.166:8080
 info:
   title: droni-api
   description: |
@@ -10,7 +12,7 @@ info:
   license:
     name: Apache License Version 2.0
     url: https://www.apache.org/licenses/LICENSE-2.0
-  version: '0.0.2'
+  version: '0.0.3'
 tags:
   - name: home
     description: home API
@@ -18,6 +20,7 @@ tags:
     description: test API
   - name: auth
     description: auth API
+  - name: article
 security:
   - bearer-key: []
 paths:
@@ -37,10 +40,6 @@ paths:
                 type: array
                 items:
                   $ref: '#/components/schemas/DroniFile'
-        404:
-          description: no banner images
-        500:
-          description: internal server error
   /home/popular-pilot:
     get:
       tags:
@@ -57,10 +56,37 @@ paths:
                 type: array
                 items:
                   $ref: '#/components/schemas/PilotProfile'
-        404:
-          description: no pilot user exist
-        500:
-          description: internal server error
+  /article/how-to-use/summary:
+    get:
+      tags:
+        - article
+      summary: 메인화면에서 보여지는 활용백서 5개 반환 api
+      responses:
+        200:
+          description: ok
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/ArticleSummaryResponse'
+  /article/drone-content/summary:
+    get:
+      tags:
+        - article
+      summary: 메인화면 드론 콘텐츠 5개 반환 api
+      responses:
+        200:
+          description: ok
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/ArticleSummaryResponse'
+
+
+
   /sucess-test:
     get:
       tags:
@@ -143,6 +169,14 @@ components:
           type: string
         refreshToken:
           type: string
+    ArticleSummaryResponse:
+      type: object
+      properties:
+        articleImageUri:
+          type: string
+        articleSubject:
+          type: string
+
 
   securitySchemes:
     bearer-key:
